fix: double entry tests
